refine delete posts function

+++ routes/post.js
@@ -62,13 +62,8 @@ router.post("/", isLoggedIn, upload2.none(), async (req, res, next) => {
 
 router.delete("/:id", isLoggedIn, async (req, res, next) => {
   try {
-    const post = await Post.findOne({ where: { id: req.params.id } });
-    if (req.user.id == post.UserId) {
-      await Post.destroy({ where: { id: req.params.id } });
-      res.send("OK");
-    } else {
-      res.send(401);
-    }
+    await Post.destroy({ where: { id: req.params.id, userId: req.user.id } });
+    res.send("OK");
   } catch (err) {
     next(err);
   }
